+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Upgrade docs dependencies.
 
+### Fixed
+
+- Coilpack drag cancel system.
+
 ## [0.3.0] 2026-05-29
 
 ### Added